+++ src/pages/HomeAdministrator.vue
@@ -1,8 +1,5 @@
 <template>
   <main class="container main">
-    <aside class="main-sidebar">
-      <AdministratorSidebar :user-info="userInfo" :actions="sidebarActions" />
-    </aside>
     <section class="main-body">
       <PvPanel header="Your administrations">
         <template #icons>
@@ -69,9 +66,7 @@ import { computed, ref, onMounted } from 'vue';
 import { storeToRefs } from 'pinia';
 import { orderByDefault, fetchDocById } from '@/helpers/query/utils';
 import { administrationCounter, administrationPageFetcher } from '../helpers/query/administrations';
-import { getSidebarActions } from '../router/sidebarActions';
 import CardAdministration from '@/components/CardAdministration.vue';
-import AdministratorSidebar from '@/components/AdministratorSidebar.vue';
 import { useAuthStore } from '@/store/auth';
 import { useQuery } from '@tanstack/vue-query';
 
@@ -83,12 +78,7 @@ const authStore = useAuthStore();
 
 const { roarfirekit, administrationQueryKeyIndex } = storeToRefs(authStore);
 
-const sidebarActions = ref(getSidebarActions(authStore.isUserSuperAdmin, false));
 
-const userInfo = ref({
-  name: 'Admin name',
-  district: 'District Name',
-});
 
 const { isLoading: isLoadingClaims, data: userClaims } = useQuery({
   queryKey: ['userClaims', authStore.uid, authStore.userClaimsQueryKeyIndex],
